Use separate entrypoint to combine compose files
Now, when you run `docker compose up`, it'll check for a `conf`
directory during start-up, and run `node install` (with a bunch of
options based on env vars). If not, it'll run `node update`.

This means that the separate `docker-compose.setup.yml` is no longer
needed, and that this plays nice with docker swarm mode.